requirement.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
arrow==0.12.1
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.22.0
AST==0.0.2
attrs==16.3.0
autobahn==18.6.1
Automat==0.5.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.3
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto==2.48.0
boto3==1.4.7
botocore==1.7.48
browsercookie==0.7.5
cachetools==2.1.0
celery==3.1.25
certifi==2018.4.16
cffi==1.11.0
channels==1.1.5
chardet==3.0.4
click==6.7
Collectfast==0.6.2
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
daphne==1.4.2
decorator==4.1.2
Django==1.11.3
django-celery==3.2.1
django-ckeditor==5.3.0
django-debug-toolbar==1.8
django-extensions==1.8.1
django-jinja==2.3.1
django-js-asset==0.1.1
django-juno-testrunner==0.4.1
django-logging-json==1.15
django-model-utils==3.0.0
django-pipeline==1.6.13
django-redis==4.8.0
django-sendsms==0.3.1
django-ses==0.8.4
django-storages==1.6.3
django-user-agents==0.3.2
django-versatileimagefield==1.8.1
django-waffle==0.12.0
django-webpush==0.2.5
djangorestframework==3.8.2
docutils==0.14
drf-extensions==0.3.1
elasticsearch==5.4.0
eliot==1.3.0
eliot-tree==17.1.0
entrypoints==0.2.3
enum34==1.1.6
Fabric==1.13.2
Flask==1.0.2
funcsigs==1.0.2
future==0.16.0
future-fstrings==0.4.3
futures==3.1.1
google-api-python-client==1.7.4
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
googlemaps==2.5.1
gunicorn==19.7.1
http-ece==1.0.5
httplib2==0.11.3
hyperlink==17.3.1
idna==2.6
imageio==2.3.0
incremental==16.10.1
ipaddress==1.0.18
ipython==5.0.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonfield==2.0.2
keyring==8.7
kombu==3.0.37
llvmlite==0.20.0
logging-tree==1.7
lxml==4.1.1
lz4==2.1.0
MarkupSafe==1.0
math-addition==3.0
moviepy==0.2.3.5
msgpack-python==0.4.8
multiprocessing==2.6.2.1
MySQL-python==1.2.5
mysqlclient==1.3.12
numba==0.35.0
numpy==1.13.3
oauth2client==4.1.2
oauthlib==2.1.0
olefile==0.45.1
pandas==0.20.3
paramiko==2.3.1
parsel==1.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py-vapid==1.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.2.0
PyNaCl==1.1.2
pyOpenSSL==17.4.0
pyrsistent==0.14.1
python-dateutil==2.6.1
python-http-client==3.1.0
python-linkedin==4.1
pytz==2017.2
pywebpush==1.6.0
queuelib==1.4.2
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
scandir==1.6
Scrapy==1.4.0
seashells==0.1.2
selenium==3.7.0
sendgrid==5.4.1
service-identity==17.0.0
shell-command==0.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sorl-thumbnail==12.4.1
SQLAlchemy==1.2.8
sqlparse==0.2.4
termcolor==1.1.0
tokenize-rt==2.0.1
toolz==0.8.2
tqdm==4.23.4
traitlets==4.3.2
tree-format==0.1.2
Twisted==17.1.0
txaio==2.10.0
ua-parser==0.7.3
unicodedata2==10.0.0.post2
uritemplate==3.0.0
urllib3==1.22
user-agents==1.1.0
w3lib==1.18.0
wcwidth==0.1.7
WebOb==1.8.1
websocket-client==0.48.0
Werkzeug==0.14.1
zope.interface==4.3.3
